server/flask-server/requirements.txt
flask
flask-bcrypt
flask_pymongo
python-dotenv
flask-session
redis
